Moved fileServer and browse.IndexPages into middleware package
